+++ tools/clang/unittests/HLSLTestLib/RDATDumper.cpp
@@ -179,8 +179,6 @@ void DumpValueArray(DumpContext &d, const char *memberName,
 }
 
 #define DEF_RDAT_TYPES DEF_RDAT_DUMP_IMPL
-#define DEF_RDAT_ENUMS DEF_RDAT_DUMP_IMPL
-#define DEF_DXIL_ENUMS DEF_RDAT_DUMP_IMPL
 #include "dxc/DxilContainer/RDAT_Macros.inl"
 
 } // namespace dump
